requirements.txt
python-telegram-bot[ext]~=21.0
gspread~=6.0
google-auth~=2.28
python-dotenv~=1.0
